# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Student-School Portal API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/schools/classes": {
            "get": {
                "tags": [
                    "Classes"
                ],
                "summary": "List all classes for the authenticated school",
                "operationId": "75d09e1344fb8cb78eb8682421c2ac91",
                "responses": {
                    "200": {
                        "description": "List of classes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "5th Grade"
                                            },
                                            "school_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "teacher_id": {
                                                "type": "integer",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Classes"
                ],
                "summary": "Create a new class for the authenticated school",
                "operationId": "6706c7841044afea347e16d6079d11cc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "5th Grade"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Class created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Class created successfully"
                                        },
                                        "class": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "5th Grade"
                                                },
                                                "school_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "teacher_id": {
                                                    "type": "integer",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/schools/classes/{class_id}": {
            "get": {
                "tags": [
                    "Classes"
                ],
                "summary": "Get details of a specific class",
                "operationId": "235aa686a03dbd260c4922c7097d44dd",
                "parameters": [
                    {
                        "name": "class_id",
                        "in": "path",
                        "description": "ID of the class",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Class details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "5th Grade"
                                        },
                                        "school_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "teacher_id": {
                                            "type": "integer",
                                            "nullable": true
                                        },
                                        "students": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Classes"
                ],
                "summary": "Update class details for the authenticated school",
                "operationId": "4be38289fdc0853b289665d40ae246c0",
                "parameters": [
                    {
                        "name": "class_id",
                        "in": "path",
                        "description": "ID of the class",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "6th Grade"
                                    },
                                    "teacher_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Class updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Class updated successfully"
                                        },
                                        "class": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "6th Grade"
                                                },
                                                "school_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "teacher_id": {
                                                    "type": "integer",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Classes"
                ],
                "summary": "Delete a class for the authenticated school",
                "operationId": "eaec35a419f178ecf1240d4c2b6f4635",
                "parameters": [
                    {
                        "name": "class_id",
                        "in": "path",
                        "description": "ID of the class",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Class deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Class deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/schools/classes/{class_id}/assign-teacher": {
            "post": {
                "tags": [
                    "Classes"
                ],
                "summary": "Assign a teacher to a class",
                "operationId": "556530da46042ca384f1ca0c0b4d72ac",
                "parameters": [
                    {
                        "name": "class_id",
                        "in": "path",
                        "description": "ID of the class",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "teacher_id"
                                ],
                                "properties": {
                                    "teacher_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Teacher assigned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Teacher assigned successfully"
                                        },
                                        "class": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "teacher_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/schools/classes/{class_id}/assign-students": {
            "post": {
                "tags": [
                    "Classes"
                ],
                "summary": "Assign students to a class",
                "operationId": "18c16e10697f6c1cb449e3817f03d022",
                "parameters": [
                    {
                        "name": "class_id",
                        "in": "path",
                        "description": "ID of the class",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "student_ids"
                                ],
                                "properties": {
                                    "student_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Students assigned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Students assigned successfully"
                                        },
                                        "class": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "students": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}